Re-implement ControllerImpl with GLFW 3.3
